requirements.txt
django==3.0.7
djangorestframework==3.11.0
Pillow==2.2.1
django-environ==0.4.5
